Move relevant logic from Entity.js to YouthActor
