+++ src/Game/entity.js
@@ -1,8 +1,6 @@
-import { ROT } from './game';
 import DynamicGlyph from './dynamicglyph';
 import EntityMixins from './entitymixins';
 import * as Messages from './messages';
-import Geometry from './geometry';
 
 // A basic entity is composed of a glyph as well as a position and a name (used in messages).
 export default class Entity extends DynamicGlyph {
@@ -117,39 +115,10 @@ export default class Entity extends DynamicGlyph {
         (target.getName() === 'sacrificed youth' ||
           target.getName() === 'sacrificed maiden')
       ) {
+        // exchange position
         this.exchangePositionWith(target);
-        // get boss direction and opposite
-        const boss = this.getMap().getBoss();
-        const direction = Geometry.getCardinal(
-          this.getX(),
-          this.getY(),
-          boss.getX(),
-          boss.getY()
-        );
-        const opposite = Geometry.getCardinal(
-          this.getX(),
-          this.getY(),
-          boss.getX(),
-          boss.getY(),
-          true
-        );
-        // create array of random messages with directions
-        const messages = [
-          `"I heard something coming from the ${direction}."`,
-          `"He's after me! I'm gonna die!"`,
-          `"I can hear it! Head ${direction} and kill the beast!"`,
-          `"I am fleeing from the beast. Head ${opposite} if you want to live!"`,
-          `"I don't want to die!"`,
-        ];
-        // Coin flip to send distress message to player
-        if (Math.round(Math.random()) === 1) {
-          Messages.sendMessage(
-            this,
-            messages[ROT.RNG.getUniformInt(0, messages.length - 1)],
-            null,
-            4
-          );
-        }
+        // raise event on target
+        target.raiseEvent('onExchange', this);
         return true;
       }
       // An entity can only attack if the entity has the Attacker mixin and
