+++ src/Game/entitymixins.js
@@ -916,6 +916,51 @@ let EntityMixins = {
       },
     },
   },
+
+  YouthActor: {
+    name: 'YouthActor',
+    groupName: 'Actor',
+    listeners: {
+      onDeath: function(attacker) {
+        // decrement remaining youths
+        this.getMap().decrementYouths();
+      },
+      onExchange: function(entity) {
+        // get boss direction and opposite
+        const boss = this.getMap().getBoss();
+        const direction = Geometry.getCardinal(
+          this.getX(),
+          this.getY(),
+          boss.getX(),
+          boss.getY()
+        );
+        const opposite = Geometry.getCardinal(
+          this.getX(),
+          this.getY(),
+          boss.getX(),
+          boss.getY(),
+          true
+        );
+        // create array of random messages with directions
+        const messages = [
+          `"I heard something coming from the ${direction}."`,
+          `"He's after me! I'm gonna die!"`,
+          `"I can hear it! Head ${direction} and kill the beast!"`,
+          `"I am fleeing from the beast. Head ${opposite} if you want to live!"`,
+          `"I don't want to die!"`,
+        ];
+        // Coin flip to send distress message to player
+        if (Math.round(Math.random()) === 1) {
+          Messages.sendMessage(
+            this.getMap().getPlayer(),
+            messages[ROT.RNG.getUniformInt(0, messages.length - 1)],
+            null,
+            4
+          );
+        }
+      },
+    },
+  },
 };
 
 // Combine TaskActor mixin with BossActor specifics
@@ -933,8 +978,6 @@ EntityMixins.BossActor = Object.assign({}, EntityMixins.TaskActor, {
       this._kills++;
       const player = this.getMap().getPlayer();
       if (prey !== player) {
-        // decrement remaining youths
-        this.getMap().decrementYouths();
         // setup a message with direction
         const direction = Geometry.getCardinal(
           player.getX(),
